Add support for building a squashed initramfs
With all files stored in ramfs, and most of them are not compressed,
the initramfs will take up a lot of memory. This becomes worse with
larger page size, because for each file, however small it is, page cache
will allocate a minimum of one page to cache the contents.

One approach to reducing the memory usage and shrink initramfs size
is to put most files into a read-only squash image and keep a minimum
set of executable and libraries outside as the loader for the squash
image. Once the squash image is mounted, the real init inside will be
executed and then everything behaves as usual.

This patch will introduce a 99squash module which will never be included
by default. User can force include it, and if it is included, dracut
will perform some extra steps before creating the final image.
"/etc" and "/usr" will be moved to "/squash", and "/init" will be renamed to
"/init.orig" and replaced by "/init.squash". Executable and libraries
required for "/init.squash" to run will be move from "/squash" to
"/squash-init", as "/squash" is not accessible before it's mounted.
"/squash" will be squashed in to "/squash.img" and "/init.sqush" will
mount it, then call "/init.orig".

According to my tests, this can save about 40MB of memory with 64K page
size, or about 20MB with 4K page size.