Add support for building a squashed initramfs
With all files stored in ramfs, and most of them are not compressed,
the initramfs will take up a lot of memory. Besides, if the file number
is large, each file will waste 1-PAGE_SIZE bytes of memory. This is due
to ramfs' design which, at least one page will be allocated for one file
however small the file is. On machine with large page size, this
will become worse and waste too many memory.

One approach to reducing the memory usage is to reduce the number of
files that got directly loaded into the root ramfs, and compress files
by put most files will into a read-only squash image and keep a minimum
set of executable and libraries outside as the loader for the squash
image. After the squash image is mounted, the real 'init' will be
executed and then everything behaves as usual.

This patch will introduce a '99squash' module which will never be
included by default. User can force add it, and if it is included,
dracut will perform some extra steps before creating the final image:

For now, "/etc" and "/usr" will be moved into the squashfs image.
"/init" will be renamed to "/init.stock" and replaced by "/init.squash".
Files and folders need to be accessible before mounting the image will
be still avaliable at their original place. And due to squashfs is
readonly, an overlayfs layer will be create on top of squashfs mount
point, as many dracut module require readwrite access to "/etc" and
"/usr".

"/init.squash" is now the init loader of the initramfs. It will do
some mounting and bind-mounting job, and call the original "/init.orig",
then everything work as usual.

This is very helpful when mem resource is very limited, like Kdump.
According to my tests, with initramfs built with mkdumprd (which is
already a smaller image than usual), this can save about 40MB of
memory with 64K page size, or about 20MB with 4K page size.

Won't change any behavior if 99squash module is not enabled.

Signed-off-by: Kairui Song <kasong@redhat.com>