squash: also squash systemctl if switch-root is not needed
systemctl need to be accessable on switch-root, but we unmount the
squash image on switch-root, so it will fail. systemctl depends on a lot
of libraries, squash them can save some memories. So allow modules (eg.
kdump) to tell dracut that switch-root will be intercepted and we don't
need to take care of that.